Basic UI for chat Fixes #10
