+++ Godot Project/Main.tscn
@@ -1,106 +1,63 @@
-[gd_scene load_steps=4 format=3 uid="uid://6g75efppoade"]
+[gd_scene load_steps=2 format=3 uid="uid://c7p4ubwm0oykt"]
 
 [ext_resource type="Script" path="res://Main.cs" id="2_i14hs"]
-[ext_resource type="Script" path="res://Match Maker/Global/Workaround.gd" id="3_8tk3s"]
-[ext_resource type="Script" path="res://Match Maker/Global/MatchMaker.cs" id="3_ucb8u"]
 
-[node name="Main" type="Node"]
-script = ExtResource("2_i14hs")
-
-[node name="MatchMaker" type="Node" parent="."]
-script = ExtResource("3_ucb8u")
-WorkaroundScript = ExtResource("3_8tk3s")
-
-[node name="Control" type="Control" parent="."]
+[node name="Main" type="Control"]
 layout_mode = 3
 anchors_preset = 15
 anchor_right = 1.0
 anchor_bottom = 1.0
 grow_horizontal = 2
 grow_vertical = 2
+script = ExtResource("2_i14hs")
 
-[node name="Panel" type="Panel" parent="Control"]
+[node name="HBoxContainer" type="HBoxContainer" parent="."]
 layout_mode = 1
 anchors_preset = 15
 anchor_right = 1.0
 anchor_bottom = 1.0
+offset_top = 32.0
+offset_bottom = 2.0
 grow_horizontal = 2
 grow_vertical = 2
 
-[node name="HBoxContainer" type="HBoxContainer" parent="Control/Panel"]
-custom_minimum_size = Vector2(0.925, 0)
-layout_mode = 1
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-
-[node name="VBoxContainer" type="VBoxContainer" parent="Control/Panel/HBoxContainer"]
-custom_minimum_size = Vector2(796, 0)
+[node name="MarginContainerPingPong" type="MarginContainer" parent="HBoxContainer"]
 layout_mode = 2
+size_flags_horizontal = 3
+theme_override_constants/margin_left = 16
+theme_override_constants/margin_top = 16
+theme_override_constants/margin_right = 16
+theme_override_constants/margin_bottom = 16
 
-[node name="PanelMatchMaker" type="Panel" parent="Control/Panel/HBoxContainer/VBoxContainer"]
-custom_minimum_size = Vector2(0, 248)
-layout_mode = 2
-
-[node name="LabelMatchMaker" type="RichTextLabel" parent="Control/Panel/HBoxContainer/VBoxContainer/PanelMatchMaker"]
+[node name="PingPongButton" type="Button" parent="HBoxContainer/MarginContainerPingPong"]
 unique_name_in_owner = true
-custom_minimum_size = Vector2(0, 192)
-layout_mode = 1
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-text = "LabelMatchMaker"
-autowrap_mode = 0
-
-[node name="PanelLocalState" type="Panel" parent="Control/Panel/HBoxContainer/VBoxContainer"]
-custom_minimum_size = Vector2(0, 192)
 layout_mode = 2
+size_flags_horizontal = 3
+text = "Ping-Pong"
 
-[node name="LabelLocalState" type="RichTextLabel" parent="Control/Panel/HBoxContainer/VBoxContainer/PanelLocalState"]
-unique_name_in_owner = true
-custom_minimum_size = Vector2(0, 192)
-layout_mode = 1
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-text = "LabelLocalState"
-autowrap_mode = 0
-
-[node name="PanelRemoteSession" type="Panel" parent="Control/Panel/HBoxContainer/VBoxContainer"]
-custom_minimum_size = Vector2(0, 192)
+[node name="MarginContainerChat" type="MarginContainer" parent="HBoxContainer"]
 layout_mode = 2
+size_flags_horizontal = 3
+theme_override_constants/margin_left = 16
+theme_override_constants/margin_top = 16
+theme_override_constants/margin_right = 16
+theme_override_constants/margin_bottom = 16
 
-[node name="LabelRemoteState" type="RichTextLabel" parent="Control/Panel/HBoxContainer/VBoxContainer/PanelRemoteSession"]
+[node name="ChatButton" type="Button" parent="HBoxContainer/MarginContainerChat"]
 unique_name_in_owner = true
-custom_minimum_size = Vector2(0, 192)
-layout_mode = 1
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-text = "LabelRemoteState"
-autowrap_mode = 0
-
-[node name="PanelMessages" type="Panel" parent="Control/Panel/HBoxContainer"]
-custom_minimum_size = Vector2(352, 0)
 layout_mode = 2
+size_flags_horizontal = 3
+text = "Chat"
 
-[node name="LabelMessages" type="RichTextLabel" parent="Control/Panel/HBoxContainer/PanelMessages"]
-unique_name_in_owner = true
-custom_minimum_size = Vector2(352, 0)
+[node name="Label" type="Label" parent="."]
 layout_mode = 1
-anchors_preset = 15
+anchors_preset = 10
 anchor_right = 1.0
-anchor_bottom = 1.0
+offset_bottom = 23.0
 grow_horizontal = 2
-grow_vertical = 2
-text = "LabelMessages"
-scroll_following = true
-autowrap_mode = 0
+theme_override_font_sizes/font_size = 32
+text = "Choose a demo:"
+horizontal_alignment = 1
+
+[connection signal="pressed" from="HBoxContainer/MarginContainerPingPong/PingPongButton" to="." method="OnPingPongButton"]
+[connection signal="pressed" from="HBoxContainer/MarginContainerChat/ChatButton" to="." method="OnChatButton"]
